requirements.txt
numpy==1.19.5
torch==1.10.0
torchvision==0.11.1
pytestf
